Urok 5

+++ src/lesson5/task1/Map.kt
@@ -206,11 +206,11 @@ fun averageStockPrice(stockPrices: List<Pair<String, Double>>): Map<String, Doub
  */
 fun findCheapestStuff(stuff: Map<String, Pair<String, Double>>, kind: String): String? {
     var res: String? = null
-    var min = Int.MAX_VALUE
+    var min = Double.MAX_VALUE
     for ((name, pair) in stuff) {
         if (pair.first == kind) {
             if (pair.second < min) {
-                min = pair.second.toInt()
+                min = pair.second
                 res = name
             }
         }
